split state into submodules
