+++ src/js/canvas/font.js
@@ -1,4 +1,4 @@
-import * as state from "../state/state.js";
+import * as state from "../state/index.js";
 import {roundToDecimal} from "../utils/numbers.js";
 
 const FONT_PT = 16; // Font size for canvas rendering
