+++ src/js/menu/file.js
@@ -1,7 +1,7 @@
 import dedent from "dedent-js";
 import Color from "@sphinxxxx/color-conversion";
 
-import * as state from "../state/state.js";
+import * as state from "../state/index.js";
 import * as actions from "../io/actions.js";
 import {defer} from "../utils/utilities.js";
 import {fontRatio} from "../canvas/font.js";
@@ -9,9 +9,9 @@ import {createDialog} from "../utils/dialogs.js";
 import exampleExportImg from "../../images/example-export.png";
 import SimpleBar from "simplebar";
 import { supported as isFileSystemAPISupported } from 'browser-fs-access';
-import * as fileSystem from "../state/file_system.js";
+import * as fileSystem from "../storage/file_system.js";
 import {ValidationError} from "../utils/errors.js";
-import {exportAnimation} from "../state/exporter.js";
+import {exportAnimation} from "../storage/exporter.js";
 import DimensionsPicker from "../components/ui/dimensions_picker.js";
 import BackgroundPicker from "../components/ui/background_picker.js";
 import UnsavedWarning from "../components/ui/unsaved_warning.js";
@@ -19,6 +19,7 @@ import {defaultContrastColor} from "../components/palette.js";
 import {modifierAbbr} from "../utils/os.js";
 import Toast from "../components/ui/toast.js";
 import {triggerRefresh} from "../index.js";
+import {DEFAULT_CONFIG} from "../state/index.js";
 
 export function init() {
     setupNew();
@@ -42,6 +43,8 @@ function setupNew() {
                 config: {
                     dimensions: [dim.numCols, dim.numRows],
                     background: backgroundPicker.value,
+                },
+                metadata: {
                     primaryColor: defaultContrastColor(backgroundPicker.value)
                 }
             })
@@ -62,8 +65,8 @@ function setupNew() {
     actions.registerAction('file.new', () => {
         unsavedWarning.toggle(state.hasCharContent());
         dimensionsPicker.value = {
-            numRows: state.CONFIG_DEFAULTS.dimensions[1],
-            numCols: state.CONFIG_DEFAULTS.dimensions[0]
+            numRows: DEFAULT_CONFIG.dimensions[1],
+            numCols: DEFAULT_CONFIG.dimensions[0]
         }
         backgroundPicker.value = false; // Transparent
         $newFileDialog.dialog('open');
@@ -117,7 +120,7 @@ const $saveFileDialog = $('#save-file-dialog');
 
 function setupSave() {
     createDialog($saveFileDialog, () => {
-        state.config('name', $saveFileDialog.find('.name').val())
+        state.setMetadata('name', $saveFileDialog.find('.name').val())
 
         fileSystem.saveFile()
             .then(() => {
@@ -179,7 +182,7 @@ function saveActive() {
                 key: 'save-active-file',
                 textCenter: true,
                 duration: 5000,
-                text: `Saved to "${state.config('name')}.${fileSystem.FILE_EXTENSION}"`
+                text: `Saved to "${state.getMetadata('name')}.${fileSystem.FILE_EXTENSION}"`
             });
         })
         .catch(err => {
@@ -249,7 +252,7 @@ function setupExport() {
 
         toggleExportPreview(format);
 
-        if (!state.config('background')) $exportOptions.find(`[name="background"]`).closest('label').hide();
+        if (!state.getConfig('background')) $exportOptions.find(`[name="background"]`).closest('label').hide();
         $('#spritesheet-png-warning').toggle(showPngSpritesheetWarning());
 
         if (EXPORT_OPTIONS[format].includes('frames')) {
@@ -330,15 +333,15 @@ function toggleExportPreview(format) {
 function openExportDialog() {
     $exportFileDialog.dialog('open');
 
-    $exportOptions.find(`[name="fps"]`).val(state.config('fps'));
+    $exportOptions.find(`[name="fps"]`).val(state.getMetadata('fps'));
 
     // If it's the first time opening the export dialog, set its values according to the last saved export settings
-    if (firstTimeOpeningExport && state.config('lastExportOptions')) {
+    if (firstTimeOpeningExport && state.getMetadata('lastExportOptions')) {
         // TODO format should be part of the options
-        $exportFormat.val(state.config('lastExportOptions').format);
+        $exportFormat.val(state.getMetadata('lastExportOptions').format);
         if (!$exportFormat.val()) $exportFormat.val($exportFormat.find('option:first').val()); // Failsafe
 
-        for (const [key, value] of Object.entries(state.config('lastExportOptions'))) {
+        for (const [key, value] of Object.entries(state.getMetadata('lastExportOptions'))) {
             const $input = $exportOptions.find(`[name="${key}"]`);
             $input.is(':checkbox') ? $input.prop('checked', !!value) : $input.val(value);
         }
@@ -406,9 +409,9 @@ function exportFromForm() {
 }
 
 function exportActive() {
-    if (!state.config('lastExportOptions')) throw new Error(`no lastExportOptions found`);
+    if (!state.getMetadata('lastExportOptions')) throw new Error(`no lastExportOptions found`);
 
-    exportAnimation(state.config('lastExportOptions'), true)
+    exportAnimation(state.getMetadata('lastExportOptions'), true)
         .then(filename => {
             new Toast({
                 key: 'export-active-file',
